% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockSVD.R
\name{blockSVD}
\alias{blockSVD}
\title{SVD using an incremental SVD algorithm.}
\usage{
blockSVD(x, ncomponents = 2, mc.cores = 2, method = "svd")
}
\arguments{
\item{x}{a real nxp matrix}

\item{ncomponents}{number of components to be computed. Default 2}

\item{mc.cores}{number of partitions of the matrix, number of svd computed in parallel.}

\item{method}{"svd" for complete svd, or "irlba" to compute only the first ncomponents.}
}
\value{
a list of two components with the singular values and left singular vectors of the matrix
}
\description{
SVD using an incremental SVD algorithm.
}
\details{
Singular values and left singular vectors of a real nxp matrix with n>p
}
\examples{
(V <- (matrix(1:30, nrow=5, ncol=6)))
blockSVD(V,k=2)
all.equal(blockSVD(V,k=2, method="svd")$u, base::svd(V)$u)
all.equal(blockSVD(V,k=2, method="irlba")$u, base::svd(V)$u[,1:2])
}
