% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalBlockSVD.R
\name{generalBlockSVD}
\alias{generalBlockSVD}
\title{SVD using an incremental SVD algorithm.}
\usage{
generalBlockSVD(A, k = 2, q = 1)
}
\arguments{
\item{k}{number of local SVDs to concatenate at each level}

\item{q}{number of levels}

\item{x}{a real nxp matrix}
}
\value{
a list of three components with the singular values and left and right singular vectors of the matrix
}
\description{
SVD using an incremental SVD algorithm.
}
\details{
Singular values and left singular vectors of a real nxp matrix
}
\examples{
(V <- (matrix(1:30, nrow=5, ncol=6)))
generalBlockSVD(V,2,3)
all.equal(generalBlockSVD(V,2,3)$u, base::svd(V)$u)
}
