% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{PCA using Jacobi algorithm or block algorithm.}
\usage{
getPCA(x, center = TRUE, scale = TRUE, method = "blockSVD",
  tol = .Machine$double.eps, ...)
}
\arguments{
\item{x}{a real nxp matrix}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance}

\item{method}{selects the method with which the function will compute the SVD. Can be: blockSVD, generalBlockSVD, Jacobi and JacobiR.}

\item{tol}{a small positive error tolerance. Default is machine tolerance.}

\item{...}{other parameters to be passed from the function of the method used.}
}
\value{
a list with the coordinates of the variables (var.coord) and the idividuals (Y), the variance of each component and its percentage and the matrix of the principal components.
}
\description{
Computation of the PCA of a matrix using different methods: Jacobi algorithm or block algorithm.
}
\details{
Principal components and variance explained for each component 
using two-sided Jacobi algorithm or block algorithm.
}
\examples{
getPCA(iris[,-5])
}
