% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jacobi.R
\name{Jacobi}
\alias{Jacobi}
\title{SVD of a symmetric matrix using Jacobi algorithm}
\usage{
Jacobi(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{a real symmetric matrix. Also data frames.}

\item{tol}{a small positive error tolerance. Default is machine tolerance}
}
\value{
a list of two components as for \code{base::eigen}
}
\description{
SVD using Jacobi algorithm
}
\details{
Eigenvalues and eigenvectores of a real symmetric matrix using 
two-sided Jacobi algorithm
}
\examples{
(V <- crossprod(matrix(1:25, 5)))
Jacobi(V)
all.equal(Jacobi(V)$values, base::eigen(V)$values)
}
