% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{offA}
\alias{offA}
\title{Auxiliar functions (not to be called by the user)}
\usage{
offA(A)
}
\arguments{
\item{A}{a real symmetric matrix}

\item{x}{tau value}

\item{V}{a real symmetrix matrix}

\item{i}{ith row}

\item{j}{jth column}

\item{ind}{vector with i and j indexes: c(i,j)}

\item{sct}{vector with (sin(theta),cos(theta),tan(theta)) of the Jacobi rotation}
}
\description{
Auxiliar functions (not to be called by the user)
}
