% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parJacobi.R
\name{parJacobi}
\alias{parJacobi}
\title{SVD using Jacobi algorithm.}
\usage{
parJacobi(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{a real symmetric matrix}

\item{tol}{a small positive error tolerance. Default is machine tolerance}
}
\value{
a list of two components as for \code{base::eigen}
}
\description{
SVD using parallel Jacobi algorithm
}
\details{
Eigenvalues and eigenvectores of a real symmetric matrix using 
two-sided Jacobi algorithm in parallel.
It needs doParallel library for parallelization.
}
\examples{
(V <- crossprod(matrix(1:25, 5)))
ncores <- detectCores() - 1
registerDoParallel(cores=ncores)
cl <- makeCluster(ncores)
parJacobi(V)
stopCluster(cl)
}
