% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgetPCA.R
\name{plot.getPCA}
\alias{plot.getPCA}
\title{Plot of the variables and individuals in PCA.}
\usage{
\method{plot}{getPCA}(x, type = "variables", comps = c(1, 2), ...)
}
\arguments{
\item{x}{an object of class getPCA}

\item{type}{variables for the variables plot or individuals for the individuals plot}

\item{comps}{components to be plot. Default c(1,2)}

\item{...}{other parameters of the plot}
}
\description{
Plots PCA
}
