% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printgetPCA.R
\name{print.getPCA}
\alias{print.getPCA}
\title{Print of the main results of the PCA performed.}
\usage{
\method{print}{getPCA}(x, ...)
}
\arguments{
\item{x}{an object of class getPCA}

\item{...}{other parameters}
}
\description{
Print PCA
}
