% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdJacobi.R
\name{svdJacobi}
\alias{svdJacobi}
\title{SVD using Jacobi algorithm}
\usage{
svdJacobi(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{a real nxp matrix}

\item{tol}{a small positive error tolerance. Default is machine tolerance}
}
\value{
a list of three components as for \code{base::svd}
}
\description{
SVD using Jacobi algorithm
}
\details{
Singular values, right singular vectors and left singular vectors of a real nxp matrix 
using two-sided Jacobi algorithm
}
\examples{
(V <- (matrix(1:30, nrow=5, ncol=6)))
svdJacobi(V)
all.equal(svdJacobi(V)$v, base::svd(V)$v)
}
