% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdJacobiR.R
\name{svdJacobiR}
\alias{svdJacobiR}
\title{SVD using Jacobi algorithm}
\usage{
svdJacobiR(x, ...)
}
\arguments{
\item{x}{a real nxp matrix}

\item{...}{other parameters to be passed from JacobiR}
}
\value{
a list of three components as for \code{base::svd}
}
\description{
SVD using Jacobi algorithm
}
\details{
Singular values, right singular vectors and left singular vectors of a real nxp matrix 
using two-sided Jacobi algorithm
}
\examples{
(V <- (matrix(1:30, nrow=5, ncol=6)))
svdJacobiR(V)
all.equal(svdJacobiR(V)$v, base::svd(V)$v)
}
